try to select different algos
