+++ CitySimulator/src/core/include/core/simulation/movement/idm/idm_movement.h
@@ -1,0 +1,12 @@
+#pragma once
+
+#include <core/simulation/movement/movement_algorithm.h>
+
+namespace tjs::core::simulation {
+	class IDMMovementAlgo : public IMovementAlgorithm {
+	public:
+		IDMMovementAlgo(TrafficSimulationSystem& system);
+
+		void update() override;
+	};
+} // namespace tjs::core::simulation
